pyproject.toml
[tool.poetry]
name = "musiccatalogue"
version = "0.1.0"
description = "Music Catalogue Api with Realtime chat"
authors = ["oaksas"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
channels = "^4.0.0"
djangorestframework = "^3.14.0"
uritemplate = "^4.1.1"
pytz = "^2024.1"
pyyaml = "^6.0.1"
sqlparse = "^0.4.4"
tzdata = "^2024.1"
django-split-settings = "^1.2.0"
flake8 = "^7.0.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pre-commit = "^3.6.2"



[tool.isort]
multi_line_output = 5
line_length = 119

[tool.yapf]
based_on_style = "google"
align_closing_bracket_with_visual_indent = true
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.project.settings"
django_find_project = false
python_files = "test_*.py"
testpaths = ["core"]
filterwarnings = "ignore::DeprecationWarning:^(?!node\\.).*:"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
